requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
Pillow
torch
torchvision
tqdm